+++ R/check.R
@@ -248,10 +248,14 @@ checkReports <- function(reports) {
   checkList(reports)
 
   for (i in seq_along(reports)) {
+    report <- reports[[i]]
     stopifnot(
-      is.character(reports[[i]]),
-      length(reports[[i]]) == 1
+      is.character(report),
+      length(report) == 1
     )
+    if (!isUrl(report) && !file.exists(report)) {
+      stop("Report must be a URL or a path to an existing file")
+    }
   }
 
   return(NULL)