+++ NEWS.md
@@ -1,4 +1,4 @@
-# 0.13.3.9000
+# 0.14.0
 
 * Support for linking to external analysis report files (`addReports()`,
 `getReports()`, `getReportLink()`). Accepts a URL or a path to file.